chore: Update Session and User models
This commit updates the Session and User models in the Prisma schema. It changes the data type of the userId field in the Session model from BigInt to Int to match the User model. Additionally, it adds a new relation between the User and Carrier models. These changes improve the consistency and integrity of the database schema.