chore: Update shop-info API endpoint
This commit adds a new API endpoint for retrieving shop information. The `api.shop-info.ts` file is created with the necessary code to authenticate the request, fetch the shop information from the Shopify API, and create a new user in the database if they don't exist. This change improves the functionality of the application by providing a way to retrieve and store shop information for future use.